requirements.txt
aiohappyeyeballs==2.4.4 ; python_version >= "3.12" and python_version < "4.0"
aiohttp-retry==2.8.3 ; python_version >= "3.12" and python_version < "4.0"
aiohttp==3.11.11 ; python_version >= "3.12" and python_version < "4.0"
aiosignal==1.3.2 ; python_version >= "3.12" and python_version < "4.0"
annotated-types==0.7.0 ; python_version >= "3.12" and python_version < "4.0"
anyio==4.8.0 ; python_version >= "3.12" and python_version < "4.0"
attrs==24.3.0 ; python_version >= "3.12" and python_version < "4.0"
cachetools==5.5.0 ; python_version >= "3.12" and python_version < "4.0"
certifi==2024.12.14 ; python_version >= "3.12" and python_version < "4.0"
cffi==1.17.1 ; python_version >= "3.12" and python_version < "4.0" and platform_python_implementation != "PyPy"
charset-normalizer==3.4.1 ; python_version >= "3.12" and python_version < "4.0"
click==8.1.8 ; python_version >= "3.12" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.12" and python_version < "4.0" and platform_system == "Windows"
cryptography==43.0.3 ; python_version >= "3.12" and python_version < "4.0"
distro==1.9.0 ; python_version >= "3.12" and python_version < "4.0"
frozenlist==1.5.0 ; python_version >= "3.12" and python_version < "4.0"
google-api-core==2.24.0 ; python_version >= "3.12" and python_version < "4.0"
google-api-python-client==2.159.0 ; python_version >= "3.12" and python_version < "4.0"
google-auth-httplib2==0.2.0 ; python_version >= "3.12" and python_version < "4.0"
google-auth-oauthlib==1.2.1 ; python_version >= "3.12" and python_version < "4.0"
google-auth==2.37.0 ; python_version >= "3.12" and python_version < "4.0"
googleapis-common-protos==1.66.0 ; python_version >= "3.12" and python_version < "4.0"
greenlet==3.1.1 ; python_version < "3.14" and (platform_machine == "aarch64" or platform_machine == "ppc64le" or platform_machine == "x86_64" or platform_machine == "amd64" or platform_machine == "AMD64" or platform_machine == "win32" or platform_machine == "WIN32") and python_version >= "3.12"
h11==0.14.0 ; python_version >= "3.12" and python_version < "4.0"
httpcore==1.0.7 ; python_version >= "3.12" and python_version < "4.0"
httplib2==0.22.0 ; python_version >= "3.12" and python_version < "4.0"
httpx==0.28.1 ; python_version >= "3.12" and python_version < "4.0"
idna==3.10 ; python_version >= "3.12" and python_version < "4.0"
jiter==0.8.2 ; python_version >= "3.12" and python_version < "4.0"
jsonpatch==1.33 ; python_version >= "3.12" and python_version < "4.0"
jsonpointer==3.0.0 ; python_version >= "3.12" and python_version < "4.0"
jsonschema-rs==0.25.1 ; python_version >= "3.12" and python_version < "4.0"
langchain-core==0.3.30 ; python_version >= "3.12" and python_version < "4.0"
langchain-openai==0.3.1 ; python_version >= "3.12" and python_version < "4.0"
langchain-text-splitters==0.3.5 ; python_version >= "3.12" and python_version < "4.0"
langchain==0.3.14 ; python_version >= "3.12" and python_version < "4.0"
langgraph-api==0.0.16 ; python_version >= "3.12" and python_version < "4.0"
langgraph-checkpoint==2.0.10 ; python_version >= "3.12" and python_version < "4.0"
langgraph-cli[inmem]==0.1.67 ; python_version >= "3.12" and python_version < "4.0"
langgraph-sdk==0.1.51 ; python_version >= "3.12" and python_version < "4.0"
langgraph==0.2.64 ; python_version >= "3.12" and python_version < "4.0"
langsmith==0.2.11 ; python_version >= "3.12" and python_version < "4.0"
msgpack==1.1.0 ; python_version >= "3.12" and python_version < "4.0"
multidict==6.1.0 ; python_version >= "3.12" and python_version < "4.0"
numpy==2.2.2 ; python_version >= "3.12" and python_version < "4.0"
oauthlib==3.2.2 ; python_version >= "3.12" and python_version < "4.0"
openai==1.59.8 ; python_version >= "3.12" and python_version < "4.0"
orjson==3.10.15 ; python_version >= "3.12" and python_version < "4.0"
packaging==24.2 ; python_version >= "3.12" and python_version < "4.0"
propcache==0.2.1 ; python_version >= "3.12" and python_version < "4.0"
proto-plus==1.25.0 ; python_version >= "3.12" and python_version < "4.0"
protobuf==5.29.3 ; python_version >= "3.12" and python_version < "4.0"
pyasn1-modules==0.4.1 ; python_version >= "3.12" and python_version < "4.0"
pyasn1==0.6.1 ; python_version >= "3.12" and python_version < "4.0"
pycparser==2.22 ; python_version >= "3.12" and python_version < "4.0" and platform_python_implementation != "PyPy"
pydantic-core==2.27.2 ; python_version >= "3.12" and python_version < "4.0"
pydantic==2.10.5 ; python_version >= "3.12" and python_version < "4.0"
pyjwt==2.10.1 ; python_version >= "3.12" and python_version < "4.0"
pyparsing==3.2.1 ; python_version >= "3.12" and python_version < "4.0"
python-dotenv==1.0.1 ; python_version >= "3.12" and python_version < "4.0"
pyyaml==6.0.2 ; python_version >= "3.12" and python_version < "4.0"
regex==2024.11.6 ; python_version >= "3.12" and python_version < "4.0"
requests-oauthlib==2.0.0 ; python_version >= "3.12" and python_version < "4.0"
requests-toolbelt==1.0.0 ; python_version >= "3.12" and python_version < "4.0"
requests==2.32.3 ; python_version >= "3.12" and python_version < "4.0"
rsa==4.9 ; python_version >= "3.12" and python_version < "4"
sniffio==1.3.1 ; python_version >= "3.12" and python_version < "4.0"
sqlalchemy==2.0.37 ; python_version >= "3.12" and python_version < "4.0"
sse-starlette==2.1.3 ; python_version >= "3.12" and python_version < "4.0"
starlette==0.45.2 ; python_version >= "3.12" and python_version < "4.0"
structlog==24.4.0 ; python_version >= "3.12" and python_version < "4.0"
tenacity==8.5.0 ; python_version >= "3.12" and python_version < "4.0"
tiktoken==0.8.0 ; python_version >= "3.12" and python_version < "4.0"
tqdm==4.67.1 ; python_version >= "3.12" and python_version < "4.0"
twilio==9.4.3 ; python_version >= "3.12" and python_version < "4.0"
typing-extensions==4.12.2 ; python_version >= "3.12" and python_version < "4.0"
uritemplate==4.1.1 ; python_version >= "3.12" and python_version < "4.0"
urllib3==2.3.0 ; python_version >= "3.12" and python_version < "4.0"
uvicorn==0.34.0 ; python_version >= "3.12" and python_version < "4.0"
watchfiles==1.0.4 ; python_version >= "3.12" and python_version < "4.0"
yarl==1.18.3 ; python_version >= "3.12" and python_version < "4.0"